requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
ecdsa==0.19.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.4
googlemaps==4.10.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
Mako==1.3.6
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.2
qrcode==8.0
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
xlrd==2.0.1
